+++ demisto_sdk/commands/common/content/objects/pack_objects/xsiam_dashboard/xsiam_dashboard.py
@@ -35,7 +35,6 @@ class XSIAMDashboard(JSONContentObject):
     def dump(self, dest_dir: Optional[Union[Path, str]] = None) -> List[Path]:
         created_files: List[Path] = []
         created_files.extend(super().dump(dest_dir=dest_dir))
-        new_file_path = str(created_files[0])
-        if 'external' in new_file_path:
-            shutil.copyfile(new_file_path, new_file_path.replace('external-', ''))
+        new_file_path = created_files[0]
+        shutil.copyfile(new_file_path, new_file_path.parent / self.normalize_file_name())
         return created_files